Adds a working example app
The original example app was missing some files
and did not run. This fixes it up. I pushed it to a
stackato VPS and purchased it on Marketplace Dev to test
it out.

After this fix, you'll want to run npm publish so that
the example dir in the receiptverifier package gets updated.